+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## 2022 Q2 Sprint 1(4/1 ~ 4/6)
 - Features
+  - [3-user-can-manage-courses](https://trello.com/c/NRBBYXNJ)
+    - `rails db:migrate`
 - Improvements
 - Bug fixes
 - Growth